+++ CHANGELOG.md
@@ -13,6 +13,10 @@
 - Fix bug preventing the iteration of multiple instances of RRule at the same time
 - Fix various bugs causing corruption of the cache in some circumstances (related to DateTime object being mutable)
 
+### Removed
+
+- The alias `RRule::occursOn` has been removed (use `occursAt` instead)
+
 ## [1.0.1] - 2016-03-11
 
 ### Fixed